+++ lib/Wikia/src/DependencyInjection/InjectorInitializer.php
@@ -3,7 +3,6 @@
 namespace Wikia\DependencyInjection;
 
 use Doctrine\Common\Cache\CacheProvider;
-use Wikia\Maintenance\UserGroups\UserGroupsModule;
 use Wikia\Service\Gateway\ConsulUrlProviderModule;
 use Wikia\Service\Swagger\ApiProviderModule;
 use Wikia\Service\User\Attributes\AttributesModule;
@@ -11,6 +10,7 @@ use Wikia\Service\User\Auth\AuthModule;
 use Wikia\Service\User\Permissions\PermissionsModule;
 use Wikia\Service\User\Preferences\Migration\PreferenceMigrationModule;
 use Wikia\Service\User\Preferences\PreferenceModule;
+use Wikia\UserGroups\UserGroupsModule;
 
 class InjectorInitializer {
 	public static function init(CacheProvider $cacheProvider = null) {
